requirements.txt
pandas
dbfread
sqlalchemy
psycopg2-binary
python-dotenv 